Update laundryScraper.py
Scrapped the idea of having a separate dynamo.py function. Instead, scraper takes the info it scrapes puts it in buildingList and machineDict, takes the info from that and puts it into buildingTable and machineTable where it is put into the proper dynamoDB format, and finally writes it to dynamo table.